fix: replace NextIP with newly added one; move Intercept func to apis utils
