+++ pkg/utils/ip_test.go
@@ -21,8 +21,6 @@ import (
 	"net"
 	"reflect"
 	"testing"
-
-	v1 "github.com/alibaba/hybridnet/pkg/apis/networking/v1"
 )
 
 func TestStringToIPNet(t *testing.T) {
@@ -69,204 +67,6 @@ func TestNormalizedIP(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestIntersect(t *testing.T) {
-	testCase := []struct {
-		name     string
-		in       []v1.AddressRange
-		expected bool
-	}{
-		{
-			"two cidr",
-			[]v1.AddressRange{
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-				},
-				{
-					Version: "4",
-					CIDR:    "192.168.100.5/24",
-				},
-			},
-			false,
-		},
-		{
-			"same cidr, no start end",
-			[]v1.AddressRange{
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-				},
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-				},
-			},
-			true,
-		},
-		{
-			"same cidr, non-overlapping start end",
-			[]v1.AddressRange{
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.10",
-					End:     "192.168.1.50",
-				},
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.51",
-					End:     "192.168.1.100",
-				},
-			},
-			false,
-		},
-		{
-			"[1]same cidr, overlap start end",
-			[]v1.AddressRange{
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.10",
-					End:     "192.168.1.50",
-				},
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.50",
-					End:     "192.168.1.100",
-				},
-			},
-			true,
-		},
-		{
-			"[2]same cidr, overlap start end",
-			[]v1.AddressRange{
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.50",
-					End:     "192.168.1.100",
-				},
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.10",
-					End:     "192.168.1.50",
-				},
-			},
-			true,
-		},
-		{
-			"[1]same cidr, overlap start end in excludedIPs",
-			[]v1.AddressRange{
-				{
-					Version:    "4",
-					CIDR:       "192.168.1.5/24",
-					Start:      "192.168.1.49",
-					End:        "192.168.1.100",
-					ExcludeIPs: []string{"192.168.1.49", "192.168.1.50"},
-				},
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.10",
-					End:     "192.168.1.50",
-				},
-			},
-			false,
-		},
-		{
-			"[2]same cidr, overlap start end in excludedIPs",
-			[]v1.AddressRange{
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.49",
-					End:     "192.168.1.100",
-				},
-				{
-					Version:    "4",
-					CIDR:       "192.168.1.5/24",
-					Start:      "192.168.1.10",
-					End:        "192.168.1.50",
-					ExcludeIPs: []string{"192.168.1.49", "192.168.1.50"},
-				},
-			},
-			false,
-		},
-		{
-			"[3]same cidr, overlap start end in excludedIPs",
-			[]v1.AddressRange{
-				{
-					Version:    "4",
-					CIDR:       "192.168.1.5/24",
-					Start:      "192.168.1.50",
-					End:        "192.168.1.100",
-					ExcludeIPs: []string{"192.168.1.50"},
-				},
-				{
-					Version:    "4",
-					CIDR:       "192.168.1.5/24",
-					Start:      "192.168.1.10",
-					End:        "192.168.1.49",
-					ExcludeIPs: []string{"192.168.1.49"},
-				},
-			},
-			false,
-		},
-		{
-			"same cidr, overlap start end not in excludedIPs",
-			[]v1.AddressRange{
-				{
-					Version:    "4",
-					CIDR:       "192.168.1.5/24",
-					Start:      "192.168.1.49",
-					End:        "192.168.1.100",
-					ExcludeIPs: []string{"192.168.1.100"},
-				},
-				{
-					Version:    "4",
-					CIDR:       "192.168.1.5/24",
-					Start:      "192.168.1.10",
-					End:        "192.168.1.50",
-					ExcludeIPs: []string{"192.168.1.10"},
-				},
-			},
-			true,
-		},
-		{
-			"same cidr, overlap start end not in excludedIPs, the same start and end",
-			[]v1.AddressRange{
-				{
-					Version:    "4",
-					CIDR:       "192.168.1.5/24",
-					Start:      "192.168.1.49",
-					End:        "192.168.1.100",
-					ExcludeIPs: []string{"192.168.1.100"},
-				},
-				{
-					Version: "4",
-					CIDR:    "192.168.1.5/24",
-					Start:   "192.168.1.50",
-					End:     "192.168.1.50",
-				},
-			},
-			true,
-		},
-	}
-
-	for _, test := range testCase {
-		t.Run(test.name, func(t *testing.T) {
-			if out := Intersect(&test.in[0], &test.in[1]); out != test.expected {
-				t.Errorf("test %s fails: expected %v but got %v", test.name, test.expected, out)
-				return
-			}
-		})
-	}
-
 }
 
 func TestValidateIP(t *testing.T) {
